requirements.txt
# requirements.txt
python-telegram-bot>=20.7
httpx>=0.24
python-dotenv>=1.0
